requirements.txt
python-docx
openai
faiss-cpu
tiktoken
langchain
